pong/requirements.txt
Django==3.2.5
psycopg2==2.8.6
djangorestframework==3.12.4
requests==2.32.3
djangorestframework_simplejwt==5.3.1
django-cors-headers==4.4.0
django-allauth==0.63.6
django-otp==1.5.2
daphne==3.0.2
channels==3.0.4
channels_redis==3.3.1
qrcode==7.4.2
python-decouple==3.8